refactor: prepare to move to selection contexts
